v2 done

+++ server/models/topic.go
@@ -4,7 +4,7 @@ import "time"
 
 // Topic is our topic model
 type Topic struct {
-	ID          uint64     `gorm:"primary_key" json:"id"`
+	ID          int        `gorm:"primary_key" json:"id"`
 	CreatedAt   time.Time  `gorm:"DEFAULT:current_timestamp" json:"created_at"`
 	UpdatedAt   time.Time  `gorm:"DEFAULT:current_timestamp" json:"updated_at"`
 	CategoryID  int        `gorm:"index" json:"category_id"`
